+++ spec/dummy/config/configs/eager_config.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class EagerConfig
+  def test
+  end
+end