+++ spec/rails/railtie_spec.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+describe "rails integrations", :rails do
+  unless ENV["DO_NOT_INITIALIZE_RAILS"] == "1" || ENV["USE_APP_CONFIGS"] == "1"
+    describe "eager_load" do
+      specify "config classes must be loaded" do
+        expect(Object.const_defined?(:EagerConfig)).to eq true
+      end
+    end
+  end
+end