+++ spec/spec_helper.rb
@@ -59,7 +59,6 @@ begin
     require "ammeter"
 
     require File.expand_path("dummy/config/environment", __dir__)
-    Rails.application.eager_load! unless ENV["DO_NOT_INITIALIZE_RAILS"] == "1"
   end
 rescue => err
   $stdout.puts "Failed to load test env: #{err}\n#{err.backtrace.take(5).join("\n")}"